Move tests into modules
