+++ geopolars/src/geoseries.rs
@@ -300,390 +300,3 @@ impl GeoSeries for Series {
         crate::ops::point::y(self)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        geoseries::{GeoSeries, GeodesicLengthMethod},
-        util::iter_geom,
-    };
-    use polars::export::arrow::array::Array;
-    use polars::prelude::Series;
-
-    use geo::{line_string, polygon, CoordsIter, Geometry, LineString, MultiPoint, Point};
-    use geozero::{CoordDimensions, ToWkb};
-    use polars::export::arrow::array::{BinaryArray, MutableBinaryArray};
-
-    use super::TransformOrigin;
-
-    #[test]
-    fn convex_hull_for_multipoint() {
-        let mut test_data = MutableBinaryArray::<i32>::with_capacity(1);
-
-        // Values borrowed from this test in geo crate: https://docs.rs/geo/0.14.2/src/geo/algorithm/convexhull.rs.html#323
-        let v = vec![
-            Point::new(0.0, 10.0),
-            Point::new(1.0, 1.0),
-            Point::new(10.0, 0.0),
-            Point::new(1.0, -1.0),
-            Point::new(0.0, -10.0),
-            Point::new(-1.0, -1.0),
-            Point::new(-10.0, 0.0),
-            Point::new(-1.0, 1.0),
-            Point::new(0.0, 10.0),
-        ];
-        let mp = MultiPoint(v);
-
-        let correct_poly: Geometry<f64> = polygon![
-            (x:0.0, y: -10.0),
-            (x:10.0, y: 0.0),
-            (x:0.0, y:10.0),
-            (x:-10.0, y:0.0),
-            (x:0.0, y:-10.0),
-        ]
-        .into();
-
-        let test_geom: Geometry<f64> = mp.into();
-        let test_wkb = test_geom.to_wkb(CoordDimensions::xy()).unwrap();
-        test_data.push(Some(test_wkb));
-
-        let test_array: BinaryArray<i32> = test_data.into();
-
-        let series =
-            Series::try_from(("geometry", Box::new(test_array) as Box<dyn Array>)).unwrap();
-        let convex_res = series.convex_hull();
-
-        assert!(
-            convex_res.is_ok(),
-            "Should get a valid result back from convex hull"
-        );
-        let convex_res = convex_res.unwrap();
-
-        assert_eq!(
-            convex_res.len(),
-            1,
-            "Should get a single result back from the series"
-        );
-        let mut geom_iter = iter_geom(&convex_res);
-        let result = geom_iter.next().unwrap();
-
-        assert_eq!(result, correct_poly, "Should get the correct convex hull");
-    }
-
-    #[test]
-    fn skew() {
-        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
-        (x: 0.0,y:0.0),
-        (x: 0.0,y:1.0),
-        (x: 1.0,y: 1.0),
-        (x: 1.0,y: 0.0)
-        ))])
-        .unwrap();
-
-        let result: Geometry<f64> = polygon!(
-            (x:-0.008727532464108793,y:-0.017460384745873865),
-            (x:0.008727532464108793,y:0.9825396152541261),
-            (x:1.008727532464109, y:1.0174603847458739),
-            (x:0.9912724675358912, y:0.017460384745873865)
-        )
-        .into();
-
-        let skewed_series = geo_series.skew(1.0, 2.0, TransformOrigin::Center);
-        assert!(skewed_series.is_ok(), "To get a series back");
-
-        let geom = iter_geom(&skewed_series.unwrap()).next().unwrap();
-
-        assert_eq!(geom, result, "the polygon should be transformed correctly");
-
-        for (p1, p2) in geom.coords_iter().zip(result.coords_iter()) {
-            assert!(
-                (p1.x - p2.x).abs() < 0.00000001,
-                "The geometries x coords to be correct to within some tollerenace"
-            );
-            assert!(
-                (p1.y - p2.y).abs() < 0.00000001,
-                "The geometries y coords to be correct to within some tollerenace"
-            );
-        }
-    }
-
-    #[test]
-    fn distance() {
-        let geo_series = Series::from_geom_vec(&[
-            Geometry::Point(Point::new(0.0, 0.0)),
-            Geometry::Point(Point::new(0.0, 0.0)),
-            Geometry::Point(Point::new(1.0, 1.0)),
-            Geometry::LineString(LineString::<f64>::from(vec![(0.0, 0.0), (0.0, 4.0)])),
-        ])
-        .unwrap();
-
-        let other_geo_series = Series::from_geom_vec(&[
-            Geometry::Point(Point::new(0.0, 1.0)),
-            Geometry::Point(Point::new(1.0, 1.0)),
-            Geometry::Point(Point::new(4.0, 5.0)),
-            Geometry::Point(Point::new(2.0, 2.0)),
-        ])
-        .unwrap();
-        let results = vec![1.0_f64, 2.0_f64.sqrt(), 5.0_f64, 2.0_f64];
-
-        let distance_series = geo_series.distance(&other_geo_series);
-        assert!(distance_series.is_ok(), "To get a series back");
-
-        let distance_series = distance_series.unwrap();
-        let distance_vec: Vec<f64> = distance_series.f64().unwrap().into_no_null_iter().collect();
-
-        for (d1, d2) in distance_vec.iter().zip(results.iter()) {
-            assert_eq!(d1, d2, "Distances differ, should be the same");
-        }
-    }
-
-    #[test]
-    fn rotate() {
-        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
-        (x: 0.0,y:0.0),
-        (x: 0.0,y:1.0),
-        (x: 1.0,y: 1.0),
-        (x: 1.0,y: 0.0)
-        ))])
-        .unwrap();
-
-        let result: Geometry<f64> = polygon!(
-        (x:0.0,y:0.0),
-        (x:-1.0,y:0.0),
-        (x:-1.0, y:1.0),
-        (x:0.0, y:1.0)
-        )
-        .into();
-
-        let rotated_series = geo_series.rotate(90.0, TransformOrigin::Point(Point::new(0.0, 0.0)));
-        assert!(rotated_series.is_ok(), "To get a series back");
-
-        let geom = iter_geom(&rotated_series.unwrap()).next().unwrap();
-        for (p1, p2) in geom.coords_iter().zip(result.coords_iter()) {
-            assert!(
-                (p1.x - p2.x).abs() < 0.00000001,
-                "The geometries x coords to be correct to within some tollerenace"
-            );
-            assert!(
-                (p1.y - p2.y).abs() < 0.00000001,
-                "The geometries y coords to be correct to within some tollerenace"
-            );
-        }
-    }
-
-    #[test]
-    fn translate() {
-        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
-        (x: 0.0,y:0.0),
-        (x: 0.0,y:1.0),
-        (x: 1.0,y: 1.0),
-        (x: 1.0,y: 0.0)
-        ))])
-        .unwrap();
-
-        let result: Geometry<f64> = polygon!(
-        (x:1.0,y:1.0),
-        (x:1.0,y:2.0),
-        (x:2.0, y:2.0),
-        (x:2.0, y:1.0)
-        )
-        .into();
-
-        let translated_series = geo_series.translate(1.0, 1.0);
-        assert!(translated_series.is_ok(), "To get a series back");
-
-        let geom = iter_geom(&translated_series.unwrap()).next().unwrap();
-        assert_eq!(geom, result, "The geom to be approprietly translated");
-    }
-
-    #[test]
-    fn scale() {
-        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
-        (x: 0.0,y:0.0),
-        (x: 0.0,y:1.0),
-        (x: 1.0,y: 1.0),
-        (x: 1.0,y: 0.0)
-        ))])
-        .unwrap();
-
-        let result_center: Geometry<f64> = polygon!(
-        (x:-0.5,y:-0.5),
-        (x:-0.5,y:1.5),
-        (x:1.5, y:1.5),
-        (x:1.5, y:-0.5)
-        )
-        .into();
-
-        let result_point: Geometry<f64> = polygon!(
-        (x:0.0,y:0.0),
-        (x:0.0,y:2.0),
-        (x:2.0, y:2.0),
-        (x:2.0, y:0.0)
-        )
-        .into();
-
-        let scaled_series = geo_series.scale(2.0, 2.0, TransformOrigin::Center);
-        assert!(scaled_series.is_ok(), "To get a series back");
-
-        let geom = iter_geom(&scaled_series.unwrap()).next().unwrap();
-        assert_eq!(
-            geom, result_center,
-            "The geom to be approprietly scaled about it's center"
-        );
-
-        let scaled_series =
-            geo_series.scale(2.0, 2.0, TransformOrigin::Point(Point::new(0.0, 0.0)));
-        let geom = iter_geom(&scaled_series.unwrap()).next().unwrap();
-        assert_eq!(
-            geom, result_point,
-            "The geom to be approprietly scaled about the point 0,0"
-        );
-    }
-
-    #[test]
-    fn euclidean_length() {
-        let mut test_data = MutableBinaryArray::<i32>::with_capacity(1);
-
-        let line_string: Geometry<f64> = line_string![
-            (x: 1., y: 1.),
-            (x: 7., y: 1.),
-            (x: 8., y: 1.),
-            (x: 9., y: 1.),
-            (x: 10., y: 1.),
-            (x: 11., y: 1.)
-        ]
-        .into();
-
-        let test_wkb = line_string.to_wkb(CoordDimensions::xy()).unwrap();
-        test_data.push(Some(test_wkb));
-
-        let test_array: BinaryArray<i32> = test_data.into();
-
-        let series =
-            Series::try_from(("geometry", Box::new(test_array) as Box<dyn Array>)).unwrap();
-        let lengths = series.euclidean_length().unwrap();
-        let as_vec: Vec<f64> = lengths.f64().unwrap().into_no_null_iter().collect();
-
-        assert_eq!(10.0_f64, as_vec[0]);
-    }
-    #[test]
-    fn explode() {
-        let point_0 = Point::new(0., 0.);
-        let point_1 = Point::new(1., 1.);
-        let point_2 = Point::new(2., 2.);
-        let point_3 = Point::new(3., 3.);
-        let point_4 = Point::new(4., 4.);
-
-        let expected_series = Series::from_geom_vec(&[
-            Geometry::Point(point_0),
-            Geometry::Point(point_1),
-            Geometry::Point(point_2),
-            Geometry::Point(point_3),
-            Geometry::Point(point_4),
-        ])
-        .unwrap();
-
-        let multipoint_0 = MultiPoint::new(vec![point_0, point_1]);
-        let multipoint_1 = MultiPoint::new(vec![point_2, point_3, point_4]);
-
-        let input_series = Series::from_geom_vec(&[
-            Geometry::MultiPoint(multipoint_0),
-            Geometry::MultiPoint(multipoint_1),
-        ])
-        .unwrap();
-
-        let output_series = GeoSeries::explode(&input_series).unwrap();
-
-        assert_eq!(output_series, expected_series);
-    }
-    #[test]
-    fn haversine_length() {
-        let mut test_data = MutableBinaryArray::<i32>::with_capacity(1);
-
-        let line_string: Geometry<f64> = LineString::<f64>::from(vec![
-            // New York City
-            (-74.006, 40.7128),
-            // London
-            (-0.1278, 51.5074),
-        ])
-        .into();
-
-        let test_wkb = line_string.to_wkb(CoordDimensions::xy()).unwrap();
-        test_data.push(Some(test_wkb));
-
-        let test_array: BinaryArray<i32> = test_data.into();
-
-        let series =
-            Series::try_from(("geometry", Box::new(test_array) as Box<dyn Array>)).unwrap();
-        let lengths = series
-            .geodesic_length(GeodesicLengthMethod::Haversine)
-            .unwrap();
-        let as_vec: Vec<f64> = lengths.f64().unwrap().into_no_null_iter().collect();
-
-        assert_eq!(
-            5_570_230., // meters
-            as_vec[0].round()
-        );
-    }
-    #[test]
-    fn vincenty_length() {
-        let mut test_data = MutableBinaryArray::<i32>::with_capacity(1);
-
-        let line_string: Geometry<f64> = LineString::<f64>::from(vec![
-            // New York City
-            (-74.006, 40.7128),
-            // London
-            (-0.1278, 51.5074),
-        ])
-        .into();
-
-        let test_wkb = line_string.to_wkb(CoordDimensions::xy()).unwrap();
-        test_data.push(Some(test_wkb));
-
-        let test_array: BinaryArray<i32> = test_data.into();
-
-        let series =
-            Series::try_from(("geometry", Box::new(test_array) as Box<dyn Array>)).unwrap();
-        let lengths = series
-            .geodesic_length(GeodesicLengthMethod::Vincenty)
-            .unwrap();
-        let as_vec: Vec<f64> = lengths.f64().unwrap().into_no_null_iter().collect();
-
-        assert_eq!(
-            5585234., // meters
-            as_vec[0].round()
-        );
-    }
-
-    #[test]
-    fn geodesic_length() {
-        let mut test_data = MutableBinaryArray::<i32>::with_capacity(1);
-
-        let line_string: Geometry<f64> = LineString::<f64>::from(vec![
-            // New York City
-            (-74.006, 40.7128),
-            // London
-            (-0.1278, 51.5074),
-            // Osaka
-            (135.5244559, 34.687455),
-        ])
-        .into();
-
-        let test_wkb = line_string.to_wkb(CoordDimensions::xy()).unwrap();
-        test_data.push(Some(test_wkb));
-
-        let test_array: BinaryArray<i32> = test_data.into();
-
-        let series =
-            Series::try_from(("geometry", Box::new(test_array) as Box<dyn Array>)).unwrap();
-        let lengths = series
-            .geodesic_length(GeodesicLengthMethod::Geodesic)
-            .unwrap();
-        let as_vec: Vec<f64> = lengths.f64().unwrap().into_no_null_iter().collect();
-
-        assert_eq!(
-            15_109_158., // meters
-            as_vec[0].round()
-        );
-    }
-}
