+++ geopolars/src/ops/affine.rs
@@ -153,3 +153,152 @@ fn translate_wkb(series: &Series, x: f64, y: f64) -> Result<Series> {
     let transform = AffineTransform::translate(x, y);
     affine_transform_wkb(series, transform)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::TransformOrigin;
+    use crate::geoseries::GeoSeries;
+    use crate::util::iter_geom;
+    use geo::{polygon, CoordsIter, Geometry, Point};
+    use polars::prelude::Series;
+
+    #[test]
+    fn rotate() {
+        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
+        (x: 0.0,y:0.0),
+        (x: 0.0,y:1.0),
+        (x: 1.0,y: 1.0),
+        (x: 1.0,y: 0.0)
+        ))])
+        .unwrap();
+
+        let result: Geometry<f64> = polygon!(
+        (x:0.0,y:0.0),
+        (x:-1.0,y:0.0),
+        (x:-1.0, y:1.0),
+        (x:0.0, y:1.0)
+        )
+        .into();
+
+        let rotated_series = geo_series.rotate(90.0, TransformOrigin::Point(Point::new(0.0, 0.0)));
+        assert!(rotated_series.is_ok(), "To get a series back");
+
+        let geom = iter_geom(&rotated_series.unwrap()).next().unwrap();
+        for (p1, p2) in geom.coords_iter().zip(result.coords_iter()) {
+            assert!(
+                (p1.x - p2.x).abs() < 0.00000001,
+                "The geometries x coords to be correct to within some tollerenace"
+            );
+            assert!(
+                (p1.y - p2.y).abs() < 0.00000001,
+                "The geometries y coords to be correct to within some tollerenace"
+            );
+        }
+    }
+
+    #[test]
+    fn scale() {
+        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
+        (x: 0.0,y:0.0),
+        (x: 0.0,y:1.0),
+        (x: 1.0,y: 1.0),
+        (x: 1.0,y: 0.0)
+        ))])
+        .unwrap();
+
+        let result_center: Geometry<f64> = polygon!(
+        (x:-0.5,y:-0.5),
+        (x:-0.5,y:1.5),
+        (x:1.5, y:1.5),
+        (x:1.5, y:-0.5)
+        )
+        .into();
+
+        let result_point: Geometry<f64> = polygon!(
+        (x:0.0,y:0.0),
+        (x:0.0,y:2.0),
+        (x:2.0, y:2.0),
+        (x:2.0, y:0.0)
+        )
+        .into();
+
+        let scaled_series = geo_series.scale(2.0, 2.0, TransformOrigin::Center);
+        assert!(scaled_series.is_ok(), "To get a series back");
+
+        let geom = iter_geom(&scaled_series.unwrap()).next().unwrap();
+        assert_eq!(
+            geom, result_center,
+            "The geom to be approprietly scaled about it's center"
+        );
+
+        let scaled_series =
+            geo_series.scale(2.0, 2.0, TransformOrigin::Point(Point::new(0.0, 0.0)));
+        let geom = iter_geom(&scaled_series.unwrap()).next().unwrap();
+        assert_eq!(
+            geom, result_point,
+            "The geom to be approprietly scaled about the point 0,0"
+        );
+    }
+
+    #[test]
+    fn skew() {
+        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
+        (x: 0.0,y:0.0),
+        (x: 0.0,y:1.0),
+        (x: 1.0,y: 1.0),
+        (x: 1.0,y: 0.0)
+        ))])
+        .unwrap();
+
+        let result: Geometry<f64> = polygon!(
+            (x:-0.008727532464108793,y:-0.017460384745873865),
+            (x:0.008727532464108793,y:0.9825396152541261),
+            (x:1.008727532464109, y:1.0174603847458739),
+            (x:0.9912724675358912, y:0.017460384745873865)
+        )
+        .into();
+
+        let skewed_series = geo_series.skew(1.0, 2.0, TransformOrigin::Center);
+        assert!(skewed_series.is_ok(), "To get a series back");
+
+        let geom = iter_geom(&skewed_series.unwrap()).next().unwrap();
+
+        assert_eq!(geom, result, "the polygon should be transformed correctly");
+
+        for (p1, p2) in geom.coords_iter().zip(result.coords_iter()) {
+            assert!(
+                (p1.x - p2.x).abs() < 0.00000001,
+                "The geometries x coords to be correct to within some tollerenace"
+            );
+            assert!(
+                (p1.y - p2.y).abs() < 0.00000001,
+                "The geometries y coords to be correct to within some tollerenace"
+            );
+        }
+    }
+
+    #[test]
+    fn translate() {
+        let geo_series = Series::from_geom_vec(&[Geometry::Polygon(polygon!(
+        (x: 0.0,y:0.0),
+        (x: 0.0,y:1.0),
+        (x: 1.0,y: 1.0),
+        (x: 1.0,y: 0.0)
+        ))])
+        .unwrap();
+
+        let result: Geometry<f64> = polygon!(
+        (x:1.0,y:1.0),
+        (x:1.0,y:2.0),
+        (x:2.0, y:2.0),
+        (x:2.0, y:1.0)
+        )
+        .into();
+
+        let translated_series = geo_series.translate(1.0, 1.0);
+        assert!(translated_series.is_ok(), "To get a series back");
+
+        let geom = iter_geom(&translated_series.unwrap()).next().unwrap();
+        assert_eq!(geom, result, "The geom to be approprietly translated");
+    }
+}
